+++ embassy-executor/CHANGELOG.md
@@ -8,14 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## Unreleased - ReleaseDate
 
+- Renamed cargo features
+  - `arch-cortex-ar` to `platform-aarch32`. 
+  - `arch-*` to `platform-*`. 
 - Added new metadata API for tasks.
 - Main task automatically gets a name of `main` when the `metadata-name` feature is enabled.
 - Upgraded rtos-trace
 - Added optional "highest priority" scheduling
 - Added optional "earliest deadline first" EDF scheduling
-- Migrate `cortex-ar` to `aarch32-cpu`. The feature name `arch-cortex-ar` remains the same and
-  legacy ARM architectures are not supported.
-- Added `run_until` to `arch-std` variant of `Executor`.
+- Migrate `cortex-ar` to `aarch32-cpu`. Noyr legacy (pre-Cortex) ARM architectures are still not supported.
+- Added `run_until` to `platform-std` variant of `Executor`.
 
 ## 0.9.1 - 2025-08-31